Async/Await Support Draft
Implement Async/Await support for WinRT with new compiler directive for
platform support. Should now create a .Net 4.5 project that also
implements Async/Await support. Not planning to add this to older
platforms as not everyone may want to take on the additional bcl
dependencies. Existing code won't use it, and new projects should be on
new platforms (if not, they can implement a wrapper themselves using
taskcompletionsource and the delegate based methods).