Stop Ignoring Tests, Fix Error Deserialisation
Took the ignore attribute off a number of tests, mostly ones that create
new tweets/dms etc. If a private test account is configured in the
config file, this should be fine.

Fixed a problem deserialising twitter error responses. It seems Twitter
now returns a collection of errors instead of a single one. Added a new
'errors' property to the response object containing the full collection,
and changed the existing 'error' property to return the first error from
the collection (if there is one).